Apply adaptable num round func
