--- app/src/main/java/com/example/traders/dialogs/BuyDialog.kt
@@ -1,78 +1,0 @@
-package com.example.traders.dialogs
-
-import android.app.AlertDialog
-import android.app.Dialog
-import android.os.Bundle
-import androidx.core.widget.addTextChangedListener
-import androidx.fragment.app.DialogFragment
-import com.example.traders.R
-import com.example.traders.databinding.FragmentBuyDialogBinding
-import com.example.traders.roundNumber
-
-class BuyDialog(val lastPrice: String, val symbol: String): DialogFragment() {
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return activity?.let {
-            val builder = AlertDialog.Builder(it)
-            val inflater =  requireActivity().layoutInflater
-
-            val view = FragmentBuyDialogBinding.inflate(inflater)
-            val dialog = builder.setView(view.root)
-                .setCancelable(true)
-                .create()
-            view.initUI()
-            view.addListeners(dialog)
-
-            dialog
-        }?: throw IllegalStateException("Activity cannot be null")
-    }
-
-    private fun enteredValIsValid(enteredVal: String): Boolean {
-        // usdAccBalance will receive val from shared prefs which are stored in activity
-        val usdAccBalance = 1525.0
-        return if(enteredVal.isBlank()) {
-            messageType = DialogValidationMessage.IS_EMPTY
-            false
-        } else if(enteredVal.toDouble() > usdAccBalance) {
-            messageType = DialogValidationMessage.IS_TOO_HIGH
-            false
-        } else if(enteredVal.toDouble() in MIN_INPUT_VAL..usdAccBalance) {
-            messageType = DialogValidationMessage.IS_VALID
-            true
-        } else if(enteredVal.toDouble() < MIN_INPUT_VAL) {
-            messageType = DialogValidationMessage.IS_TOO_LOW
-            false
-        } else {
-            messageType = DialogValidationMessage.IS_TOO_HIGH
-            false
-        }
-    }
-
-    companion object {
-        val MIN_INPUT_VAL = 10.0
-        var messageType = DialogValidationMessage.IS_VALID
-    }
-
-    private fun FragmentBuyDialogBinding.initUI() {
-        cryptoPriceLabel.text = cryptoPriceLabel.context.getString(R.string.price_of_coin, symbol)
-        cryptoAmountLabel.text = cryptoAmountLabel.context.getString(R.string.coin_amount, symbol)
-        cryptoPrice.text = "$ " + roundNumber(lastPrice.toDouble())
-    }
-
-    private fun FragmentBuyDialogBinding.addListeners(dialog: AlertDialog) {
-        priceInputField.addTextChangedListener { enteredVal ->
-            buyBtn.isEnabled = enteredValIsValid(enteredVal.toString())
-            validationMessage.text = messageType.message
-        }
-
-        cancelBtn.setOnClickListener {
-            dialog.dismiss()
-        }
-
-        buyBtn.setOnClickListener {
-            dialog.dismiss()
-        }
-    }
-}
-
-
